refactor: rename ExpensePageList to ExpenseList for consistency
- Remove unused old ExpenseList component
- Rename ExpensePageList to ExpenseList to match naming pattern
- Update component name, interface, and imports
- Fix lotId type conversion in ContributionModal (number to string)
- Maintain consistent naming with IncomeList and LotList components
